requirements.txt
python-telegram-bot==20.7
moviepy==1.0.3
python-dotenv==1.0.0
imageio==2.31.1
imageio-ffmpeg==0.4.8
